Minor reorganization of shader validation code
